requirements.txt
# For EXIF extraction in picture_utils.py
piexif
# For testing (used by test_server.py etc)
httpx
open_clip_torch
fastapi
uvicorn
setuptools
wheel
twine
numpy
pillow
opencv-python
scipy
platformdirs
tomli
colorlog
onnxruntime
requests
transformers
torch
torchvision
insightface
rapidfuzz
tqdm
einops
spacy
sentence_transformers
pillow-heif
sqlmodel
nltk
